+++ packages/xss/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.16 (December 20, 2024)
+
+- chore: update deps
+
 ## 0.1.15 (December 11, 2024)
 
 - chore: update deps